% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{add_commas}
\alias{add_commas}
\title{Add comma formatting to DT::datatable columns}
\usage{
add_commas(table, columns = c("Raw SKUs"))
}
\arguments{
\item{table}{DT::datatable}

\item{columns}{A vector with column names}
}
\value{
Formatted DT::datatable
}
\description{
Add comma formatting to DT::datatable columns
}
\examples{
\dontrun{table \%>\% add_commas("n_sku")}
}
