% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_gb_functions.R
\name{empty_to_na}
\alias{empty_to_na}
\title{Convert empty string values "" to NA}
\usage{
empty_to_na(input)
}
\arguments{
\item{input}{Text field}
}
\value{
Text field
}
\description{
Convert empty string values "" to NA
}
\examples{
\dontrun{df \%>\% mutate(myvar = empty_to_na(myvar))}

\dontrun{df \%>\% mutate_if(is.character, funs(ifelse(. == "", NA, .)))}
}
