% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{get_extended_tree}
\alias{get_extended_tree}
\title{Add industry and sector columns to Passport industry tree}
\usage{
get_extended_tree(df)
}
\arguments{
\item{df}{data frame with Euromonitor Passport Tree with the following columns: categoryid, productname, categoryparentid, indentlevel, projectid, islowestlevel}
}
\value{
data frame with the following extra columns: industryid, industry, sectorlev1id, sectorlev1
}
\description{
Add industry and sector columns to Passport industry tree
}
\examples{
\dontrun{df \%>\% get_extended_tree()}
}
