% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_gb_functions.R
\name{\%+ana\%}
\alias{\%+ana\%}
\title{Special average}
\usage{
x \%+ana\% y
}
\arguments{
\item{x}{numeric field}

\item{y}{numeric}
}
\value{
numeric
}
\description{
Take an average of variables which are nonmissing
}
\examples{
x \%+ana\% y
}
