% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_gb_functions.R
\name{\%+na\%}
\alias{\%+na\%}
\title{Special sum}
\usage{
x \%+na\% y
}
\arguments{
\item{x}{numeric field}

\item{y}{numeric}
}
\value{
numeric
}
\description{
Add variables which are nonmissing
}
\examples{
x \%+na\% y
}
