% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_gb_functions.R
\name{plot_table_btns}
\alias{plot_table_btns}
\title{Create a DT::datatable with export data buttons and nice other defaults}
\usage{
plot_table_btns(df, page_length = 20)
}
\arguments{
\item{df}{Final table in a data frame format}

\item{page_length}{Number of rows to show at a time}
}
\value{
Formatted DT::datatable
}
\description{
Create a DT::datatable with export data buttons and nice other defaults
}
\examples{
\dontrun{df \%>\% plot_table_btns()}
}
