% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_gb_functions.R
\name{sumstats}
\alias{sumstats}
\title{Create a nice summary table with number of observations, and other statitics}
\usage{
sumstats(df, detailed = FALSE, text_len = 6)
}
\arguments{
\item{df}{data frame}

\item{detailed}{should we show additional statistics default is FALSE}

\item{text_len}{the length of text in min and max calculation, default is 6}
}
\value{
numeric
}
\description{
Create a nice summary table with number of observations, and other statitics
}
\examples{
sumstats(df, TRUE, 12)
}
