% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_gb_functions.R
\name{trim}
\alias{trim}
\title{Remove leading or trailing whitespace}
\usage{
trim(x)
}
\arguments{
\item{x}{Text field}
}
\value{
Text field
}
\description{
Remove leading or trailing whitespace
}
\examples{
\dontrun{df \%>\% mutate(x = trim(x))}
}
