% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_gb_functions.R
\name{empty_to_na}
\alias{empty_to_na}
\title{Convert empty string values "" to NA}
\usage{
empty_to_na(input)
}
\arguments{
\item{input}{Text field}
}
\value{
Text field
}
\description{
Convert empty string values "" to NA
}
\examples{
load(system.file("extdata", "string_vector.rda", package = "econmodme"))
empty_to_na(string_vector)

\dontrun{df \%>\% mutate_if(is.character, funs(ifelse(. == "", NA, .)))}
}
