% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_gb_functions.R
\name{remove_special_symbols}
\alias{remove_special_symbols}
\title{Remove special symbols}
\usage{
remove_special_symbols(x)
}
\arguments{
\item{x}{Text field}
}
\value{
Text field
}
\description{
Remove special symbols
}
\examples{
\dontrun{df \%>\% mutate(x = remove_special_symbols(x))}
}
