% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_gb_functions.R
\name{trim}
\alias{trim}
\title{Remove leading or trailing whitespace}
\usage{
trim(x)
}
\arguments{
\item{x}{Text field}
}
\value{
Text field
}
\description{
Remove leading or trailing whitespace
}
\examples{
load(system.file("extdata", "string_vector.rda", package = "econmodme"))
trim(string_vector)
}
